Return `gb18030` instead of `gbk` (#149)
* Return `gb18030` instead of `gbk`

Chardetng doesn't differentiate between `gbk` and `gb18030` encodings because it is designed to work with encoding_rs, which decodes both of them the same. However, Python's built-in decoders are more strict, and if you try to decode some `gb18030` bytes with the `gbk` decoder, it'll raise an exception. The other way around is fine, though, since `gb18030` is a superset of `gbk`. To make things simpler for users, this translates Chardetng's `gbk` to `gb18030` when passing results back to Python code.

* Do lower-case comparisons on encoding names